Split the shifts.ts into shifts and assignments.
