+++ collectivo/extensions/shifts/composables/assignments.ts
@@ -1,0 +1,74 @@
+import { DateTime } from "luxon";
+import {
+  getNextOccurrence,
+  isShiftDurationModelActive,
+} from "~/composables/shifts";
+
+export const isNextOccurrenceWithinAssignment = (
+  assignment: ShiftsAssignment,
+): boolean => {
+  if (typeof assignment.shifts_slot == "number") {
+    throw new Error("assignment.shifts_slot field must be loaded");
+  }
+
+  if (typeof assignment.shifts_slot.shifts_shift == "number") {
+    throw new Error("assignment.shifts_slot.shifts_shift field must be loaded");
+  }
+
+  const nextOccurrence = getNextOccurrence(assignment.shifts_slot.shifts_shift);
+
+  return (
+    !!nextOccurrence &&
+    isShiftDurationModelActive(assignment, nextOccurrence.start)
+  );
+};
+
+export const getActiveAssignment = (
+  assignments: ShiftsAssignment[],
+  atDate?: DateTime,
+): ShiftsAssignment | null => {
+  for (const assignment of assignments) {
+    if (isShiftDurationModelActive(assignment, atDate)) return assignment;
+  }
+
+  return null;
+};
+
+export const getAssigneeName = (
+  assignments: ShiftsAssignment[],
+  atDate?: DateTime,
+) => {
+  if (!atDate) {
+    atDate = DateTime.now();
+  }
+
+  const assignment = getActiveAssignment(assignments, atDate);
+
+  if (!assignment)
+    return "No assignee on " + atDate.toLocaleString(DateTime.DATE_SHORT);
+
+  if (typeof assignment.shifts_user == "string") {
+    throw new Error("Assignment shifts_user field must be loaded");
+  }
+
+  return (
+    assignment.shifts_user.first_name +
+    " " +
+    assignment.shifts_user.last_name[0] +
+    "."
+  );
+};
+
+export const hasActivePermanentAssignment = (
+  assignments: ShiftsAssignment[],
+  atDate?: DateTime,
+) => {
+  if (!atDate) {
+    atDate = DateTime.now();
+  }
+
+  const assignment = getActiveAssignment(assignments, atDate);
+  if (!assignment) return false;
+
+  return assignment.shifts_to == undefined;
+};
